Fix `UNIT_DIED` logic to remove last video segment
When `UNIT_DIED` is logged, save the time of the log line and use it
when the Challenge Mode dungeons ends to determine if we should remove
the last video segment or not.

There will _always_ be a trash video segment at the end, unless we do
this, and in most cases there will be no trash killed after the
encounter ends.

Often, the `UNIT_DIED` line for the boss in the encounter will happen a
few milliseconds after the `ENCOUNTER_END` which is why this logic is
needed.