Fix memory leak from EventEmitter
Fixes #86.

The listener on `ipc.on('refreshState')` was called on every re-render,
and this commit fixes that by making sure the listener only attached
once per component load/unload.

This caused memory leaks as logged in the console:

   (node:14384) MaxListenersExceededWarning: Possible EventEmitter memory leak detected.
   11 refreshState listeners added to [EventEmitter]. Use emitter.setMaxListeners() to increase limit

`React.useEffect()` triggers on every state update, but when given an
empty array as its 2nd argument, it changes that to only trigger on
component load/unload.

The scenario is described in the React documentation here:

https://www.robinwieruch.de/react-function-component/#react-function-component-lifecycle

under section "React Function Component: Lifecycle".