Improve `handleLogLine` and `splitLogLine`
`splitLogLine` no longer assumes a timestamp length of max. 18
characters, but instead looks for the first two spaces in the line
string and starts parsing from there, but still saving the timestamp
part to avoid re-parsing later.

This is more in line with how the old simple `.split(',')` method was,
as it would have the first element contain the date (as well as the
line type token).

The method now has a `maxSplits` argument that, when given, will limit
the maximum number of parsed arguments returned. Technically, it's
advisable to only parse up to as many arguments as needed, but that
isn't done yet as that'd probably be a pre-emptive overoptimization.

The method now returns an instance of the new class `LogLine`, which
contains both the timestamp and the line arguments. At the same time,
all `handle*Line` methods have been updated to accept an argument of
`LogLine` instead of a string.

`getCombatLogDate` has been moved to `LogLine` and simply called
`date()`, all occurances of the former has been replaced by the latter.

This is not so much a performance improvement more than it's a logical
grouping of methods/data.

`handleLogLine` now checks the line token type, like `COMBATANT_INFO`
against an array of `interestingCombatLogEvents` and will only parse the
full log line if one matches.