Implement the visuals of the audio device selection
To ease the saving of future settings, an array has been created with
the settings to save and `saveSettings` will simply save all of them
when called.

Fetches a list of input/output audio devices via the new
`getAudioDevices` IPC channel and creates arrays with options for
`<select />` elements to pick from, including options for 'All' and
'None'.