Fix audio sources not being removed
When selecting 'None' for input or output audio devices, it'll work the
first time the app is started, but if you select anything else and then
go back to 'None' it won't work because the method to do this was to
simply ignore the devices we didn't want to record.

For some reason, they linger in the OBS configuration so a fix around
this is to configure all sources, as previously done, and instead use
the `muted` flag on the source to select whether to record it or not.