Change how the last video segment is treated
Instead of tracking `UNIT_DIED` to check if something was killed after
the last encounter, we'll simply check if the last segment is too short
to be meaningful (10 seconds).

Most often, the important parts of a dungeon isn't in the end after the
last boss anyway.

Also, fix `handleChallengeModeStart()` not working when doing multiple
dungeons in one go - it was due to `activeChallengeMode` not being
cleared after one ends. Instead of clearing it just assume we won't have
one staring while another one is in progress - logically, it shouldn't
happen.