Replace `string.split()` with `splitLogLine()`
`splitLogLine` changes:

Split log lines intelligently with respect to quotes, lists, tuples and
what have we such that we can better use the data contained within the
log lines.

`splitLogLine`  has a `maxSplits` argument that, when given, will limit
the maximum number of parsed arguments returned. Technically, it's
advisable to only parse up to as many arguments as needed, but that
isn't done yet as that'd probably be a pre-emptive overoptimization.

The method now returns an instance of the new class `LogLine`, which
contains both the timestamp and the line arguments. At the same time,
all `handle*Line` methods have been updated to accept an argument of
`LogLine` instead of a string.

`getCombatLogDate` has been moved to `LogLine` and simply called
`date()`, all occurances of the former has been replaced by the latter.

This is not so much a performance improvement more than it's a logical
grouping of methods/data.

`handleLogLine` changes:

`handleLogLine` has been refactored to use a `switch..case` style matching
to avoid doing text searching with `string.contains()` for every line,
which in a worst case scenario would perform this check as many times
as there are `if`-conditions, which would be most of the time since we're
only interested in a very limited subset of the combat log events that
WoW logs.

`handleLogLine` now checks the line token type, like `COMBATANT_INFO`
against an array of `interestingCombatLogEvents` and will only parse the
full log line if one matches to improve the processing speed of combat
events.

`handleRaid(Start|Stop)Line` has been renamed to `handleEncounter(Start|Stop)Line`
as `ENCOUNTER_(START|STOP)` isn't exclusive to raids.