Replace `string.split()` with `splitLogLine()` (#95)
* Replace `string.split()` with `splitLogLine()`

`splitLogLine` changes:

Split log lines intelligently with respect to quotes, lists, tuples and
what have we such that we can better use the data contained within the
log lines.

`splitLogLine`  has a `maxSplits` argument that, when given, will limit
the maximum number of parsed arguments returned. Technically, it's
advisable to only parse up to as many arguments as needed, but that
isn't done yet as that'd probably be a pre-emptive overoptimization.

The method now returns an instance of the new class `LogLine`, which
contains both the timestamp and the line arguments. At the same time,
all `handle*Line` methods have been updated to accept an argument of
`LogLine` instead of a string.

`getCombatLogDate` has been moved to `LogLine` and simply called
`date()`, all occurances of the former has been replaced by the latter.

This is not so much a performance improvement more than it's a logical
grouping of methods/data.

`handleLogLine` changes:

`handleLogLine` has been refactored to use a `switch..case` style matching
to avoid doing text searching with `string.contains()` for every line,
which in a worst case scenario would perform this check as many times
as there are `if`-conditions, which would be most of the time since we're
only interested in a very limited subset of the combat log events that
WoW logs.

`handleLogLine` now checks the line token type, like `COMBATANT_INFO`
against an array of `interestingCombatLogEvents` and will only parse the
full log line if one matches to improve the processing speed of combat
events.

`handleRaid(Start|Stop)Line` has been renamed to `handleEncounter(Start|Stop)Line`
as `ENCOUNTER_(START|STOP)` isn't exclusive to raids.

* Make `LogLine.date()` tolerate no day/month

Some of the tests in `resources/test-scripts/` don't have day/month for
the timestamps, which threw `date()` off and caused a `null` value date,
which in turned crashed ffmpeg due to a video duration of `NaN`.

* Address comments in PR #95

- `LogLine.ts` renamed to `timestamp` to avoid confusing it with
  `TypeScript`.

- `splitLogLine()` has new comments that clarify the meaning of the
  arguments as well as some of the array variables used in the method.

- Handle `CHALLENGE_MODE_(START|END)` as a guard to ignore
  `ENCOUNTER_(START|END)` during Mythic Keystone dungeons that aren't
  supported yet to avoid recording boss kills for a dungeon.

- Refactor `handleLogLine` logic to avoid duplicating the combat log
  events names, that we're interested in handling.

  Basically, a function signature has been defined
  (`LogLineHandlerType`), which is how the handler functions needs to be
  defined, then `interestingCombatLogEvents` has been changed to an
  object with the keys being the names of the functions that handle each
  combat log item.

  They are called with the parsed log line, as previously, but with more
  magic.

- Refactor all `handle*Line()` expressions to proper `function`s due to
  JavaScripts function hoisting which doesn't work when it's defined as
  a `const` expression, but allows us to "use" the function name (in
  `interestingCombatEvents`) before it's actually defineactually
  defined.

- Remove misleading commend about 'tuples' in `splitLogLine()`