Feature: Select audio device (#89)
* Implement selection of audio devices in settings

Due to how the renderer doesn't have access to `NodeJS.process`
(for `process.platform`), the audio device discovery, by means of
`getAudioDevices` has been moved from `obsRecorder.ts` to
`obsAudioDeviceUtils.ts`.

Two new convenience functions have been created, that does exactly what
they say:

  - `getAvailableAudioInputDevices()`
  - `getAvailableAudioOutputDevices()`

`getAudioDevics` now return an array of `ObsAudioDevice` instances, for
better typing in TypeScript across modules. OBS mostly return "unknown"
or "any" types.

In `setupSources` in `obsRecorder.ts`, we now load two new settings and
filter audio devices based on those settings as such:

- 'none' = no devices included
- 'all' = all devices included
- '<device id>' = single device as selected

* Ensure audio devices are released when changing

When selecting a different audio device, we'll need to make sure that
the devices are properly released after the call to `osn.InputFactory.create()`
otherwise we might leave a Bluetooth headphone in a low bitrate
"headset" state.

* Add an IPC channel for fetching audio devices

The renderer cannot directly use these methods because of the lack of
access to `NodeJS.process` in the browser, so we'll use an IPC channel.

* Implement the visuals of the audio device selection

To ease the saving of future settings, an array has been created with
the settings to save and `saveSettings` will simply save all of them
when called.

Fetches a list of input/output audio devices via the new
`getAudioDevices` IPC channel and creates arrays with options for
`<select />` elements to pick from, including options for 'All' and
'None'.

* Add changelog entry for the feature

* Fix getting config the right way

Adresses the issue of not retrieving config settings the right way.

* Increase size of Settings window

To accomodate the ever increasing number of settings.

* Fix React errors in Settings

- `<option />` elements each have a unique `key` property now

- `<select />` now uses `value` property to show the selected value

* Create `getStringConfigSafe()` for string values

This streamlines the config retrieval interface, which uses
`get<something>ConfigSafe()`.

Also, change `/` in `getPathConfigSafe()` to `path.sep` which uses a
platform specific separator instead of a hardcoded one.

* Typehint `cfg` as `ElectronStore`

This makes it easier to verify the types and such during development

* Resolve a mixup of input/output devices

There was a bad mixup of input and output devices whch meant that we'd
never be able to record any output audio.

This has been fixed.

* Fixed update error and simplified settings

- Replaced the numerous `React.useState()` variables with a single
  instance of `React.useState()`, as seen in fx. `Layout.tsx`

- Combined and simplified the `update*Value()` callbacks for the input
  fields into the `setSetting()` function, which now takes a state key
  (rather than the store key) and a value and updates both the `value`
  property of the appropriate input element as well as the state.

- Fixed `audio*Device` state not being updated on selection and
  therefore not updating the selected item in the `<select>` list.

* Fix audio sources not being removed

When selecting 'None' for input or output audio devices, it'll work the
first time the app is started, but if you select anything else and then
go back to 'None' it won't work because the method to do this was to
simply ignore the devices we didn't want to record.

For some reason, they linger in the OBS configuration so a fix around
this is to configure all sources, as previously done, and instead use
the `muted` flag on the source to select whether to record it or not.