Make video player settings global
When a video is muted or the volume changed, that setting would only
persist within the selected video category (e.g. 3v3) and if a different
category was selected, the settings would reset which was undesirable.

This has been changed such that the settings are _global_ across the app
regardless of category and video selection.

In types.ts:

- New type `VideoPlayerSettings` which contains the muted/volume state
  values that we want to preserve.

In preload.ts:

- New IPC channel `videoPlayerSettings` with two main functions as
  args[0]:

  'get': retrieve the stored settings
  'set': store the settings given in args[1]

In main.ts:

- IPC listener to get/set video player settings.

In Layout.tsx:

- New function `handleVideoPlayerVolumeChange()` which handles saving
  the state of the muted/volume when it changes and sets the values in
  the existing state, but doesn't apply it as that will happen in due
  time whenever a new video or category is selected.