Fix crashing with invalid configuration
Fixes issue #119 which describes a problem in which an invalid
configuration, as how the initial state of the app is for new users,
will crash when one attempts to open the Settings window.

This is due to OBS not yet being initialized but attempted to be used
to retrieve available input/output audio devices.