Refactor/rewrite the combat parser logic
The combat log parser, watcher, tail'er, and what have we has been
refactored and some rewritten to support monitoring multiple log files
simultaneously.

- `LogLine.parseLogArg()` has had its `continue` statements in the
  `switch (c)` block replaced by `break`, followed by a single
  `continue` at the end because TypeScript considered `continue`
  as a fall-through.

- `LogLine` has a new method, `addArg()` that simply facilitates adding
  a parsed argument to `this.args`.

New `CombatLogParser` class that replaces all previous logic related to
the parsing of, discovery of, and tailing of combat log files that was
in `logutils.ts`.

This class is instantiated with a single option, a data timeout. This
timeout is used to fire an event if no data has been received from the
active combat log, which can be used to stop recording or inform the
user.

The class extends `EventEmitter` and therefore emits combat log events
as JavaScript events as we know them best. All combat log handlers are
now attached as event listeners using `.on('....', callback)` as seen in
`logutils.ts`. All event listerners will receive the full `LogLine` as
well as a 2nd argument, `flavour`, which will indicate which flavour of
WoW the combatlog was parsed from.

Whenever data is received on a specific combat log, that log will have
exclusive data processing for up to `dataTimeout` seconds such that the
recorder will not receive events from two combat logs from differente
flavours of WoW at the same time.