Add a button to manually stop recording
Sometimes, the app doesn't stop recording on its own for various reasons
so this change introduces a way to stop recording manually.

Whenever the app is recording and the "REC" icon is flashing, clicking
on that icon will change it to a stop icon and clicking it will ask for
confirmation to do so.

In the event of a force stop, the following will happen to attempt to
clean up:

  - Current recording `duration` is set to `<current time> - videoStartDate`
  - Combatant map and player is cleared
  - Any active challenge mode object is cleared
  - Any active test is cleared
  - Recorder will be stopped with whatever current metadata is present

`runRecordingTest` has been modified as such:

  - It now allows for testing the above functionality by holding down
    Ctrl and Alt while clicking the test button. This causes the test to
    never end on its own and requires the use of the force stop button.

  - Timestamp for log lines in this function wil be set on-the-fly
    to match current date and time + an optional adjustment in seconds
    to avoid confusing the app when the start date is too far in the
    past.

App CSS has been changed slightly to accomodate the new
`<ConfirmationDialog/>` component:

  - `h2` is no longer styled for `.MuiDialogTitle-root` elements
  - `.MuiDialogTitle-root` elements no longer has the styling from
    `.MuiTypography-root`.

This was to make it looks not great, but also not terrible.