Fix UNIT_DIED breaking before recording starts
In the case of raids, it's not unusual for a player to die _before_ an
encounter starts for various reasons (e.g. trash) but as we don't have a
`videoStartDate` before the recording/encounter starts, this caused the
application to crash due to trying to access `videStartDate.getTime()`.

This fix simply guards `handleUnitDiedLine()` in such a way where if we
are not recording or if we don't have a valid `videoStartdate`, we'll
ignore `UNIT_DIED`.

Fixes issue #134