Name video files more sensibly
Introduced more constants for raids, which adds an easy and scalable way
to find a raid name from an encounter ID; `raidInstances` is an array of
raids with zone ID, name, and a list of encounters (by encounter ID).

Hashmap `instanceEncountersById` has been added to have a single point
of truth regarding encounter IDs to encounter name for all instances
that support this (dungeons, raids).

A new file has been added, `helpers.ts`, which should be kept FREE from
any filesystem related modules (as these cannot be imported in the
renderer process) and some of the helper functions from both `utils.ts`
as well as `renderutils.ts` has been moved here, as they were needed
in multiple places.

This file utilizes the format of `export const methodName = () => {};`
as it avoids having a long list of exported names at the bottom, and
makes it easy to identify which methods are exported and which aren't.

Furthermore, it vastly improves renaming symbols across the entire
codebase because it isn't just changed to `export { newName as oldName
}`.

The `Recorder` class will now generate a suitable filename for the final
video file based on the given metadata to name the final video files a
bit more sensibly.

The format that was chosen results in file names like:

  - `2022-09-22 15-36-33 - {2v2} Enigma Crucible (Win)`
  - `2022-09-22 16-15-54 - {Raids} Sepulcher of the First Ones, Lihuvim (Wipe)`
  - `2022-09-22 15-44-19 - {Mythic+} Mechagon Junkyard +18 (+1)`

So, basically: `{timestamp} - {category} {Zone/Instance}` + some
category specific information as seen above, whenever possible.

Replaced all occurances of `{ [id: number]: string; }` with
`NumberKeyToStringValueMapType` in `constants.ts`.