Make buffer storage path configurable
`bufferStoragePath` defaults to the value of `${storageDir}/.temp` in
the Settings window, but this is only visual - it doesn't save that
value anywhere. It's just to show people that it doesn't need to be set.

A guard has been put in place in `saveSettings()` to PREVENT Settings
from being saved if `storagePath` and `bufferStoragePath` have the same
value as this is an invalid configuration and would lead to LOSS of
video files when `cleanupBuffer()` runs.

Merged `openStoragePathDialog()` and `openLogPathDialog()` into a new
method, `openDirectorySelectorDialog()`, which is generic and takes a
settings key for the path we're selecting.

Renamed `ConfirmationDialog` => `InformationDialog` and made it a
generic dialog that can be used to display whatever type of
information, confirmation, question, and what have we.