Ensure we save all settings correctly
`getAttribute()` returns `null`, specifically, when an element does not
have the given attribute but the logic wasa simply testing the `value`
as being 'truthy', which meant that `0` or even an empty string wouldn't
be saved.

Monitor indexes are zero based and therefore we couldn't save monitor
index '0'. We can now!