Make finding resolutions more robust
The previous method of checking if a number was close to another number
was a bit fuzzy.

This change introduces a new method `getClosestResolution()` which,
given an array of available resolutions from OBS and a target resolution
to match, will return the one that _closest_ match.

This means we'll _never_ crash due to not being able to find a
resolution that matches, as it'll always return the one that closest
match.

`isNumberClose()` has been removed due to this.