Refactor: Log parsing performance & multiple log paths (#156)
* Merge `LogLine` and `splitLogLine`

Refactored `LogLine` and `splitLogLine()` for performance reasons, to
avoid parsing arguments of a combat log line that we're not interested
in.

Many combat log events have arguments that we don't use and there's no
point in parsing them all.

This change primarily ensures that we parse on-demand.

Notable changes:

- `LogLine.args[n]` has been replaced by `LogLine.arg(n)`

- `LogLine` now takes the original combat line as its only constructor
  argument.

- `handleLogLine()` no longer calls `splitLogLine()` twice since that
  isn't necessary anymore.

- Renamed variables in `splitLogLine()` from snake_case to
  lowerCamelCase as the rest of the codebase.

* Remove trailing white-space from all lines

* Refactor/rewrite the combat parser logic

The combat log parser, watcher, tail'er, and what have we has been
refactored and some rewritten to support monitoring multiple log files
simultaneously.

- `LogLine.parseLogArg()` has had its `continue` statements in the
  `switch (c)` block replaced by `break`, followed by a single
  `continue` at the end because TypeScript considered `continue`
  as a fall-through.

- `LogLine` has a new method, `addArg()` that simply facilitates adding
  a parsed argument to `this.args`.

New `CombatLogParser` class that replaces all previous logic related to
the parsing of, discovery of, and tailing of combat log files that was
in `logutils.ts`.

This class is instantiated with a single option, a data timeout. This
timeout is used to fire an event if no data has been received from the
active combat log, which can be used to stop recording or inform the
user.

The class extends `EventEmitter` and therefore emits combat log events
as JavaScript events as we know them best. All combat log handlers are
now attached as event listeners using `.on('....', callback)` as seen in
`logutils.ts`. All event listerners will receive the full `LogLine` as
well as a 2nd argument, `flavour`, which will indicate which flavour of
WoW the combatlog was parsed from.

Whenever data is received on a specific combat log, that log will have
exclusive data processing for up to `dataTimeout` seconds such that the
recorder will not receive events from two combat logs from differente
flavours of WoW at the same time.

* Add all 3 log flavours to Settings

It doesn't look great, but it works for now until it can be redone.

* Address comments in PR #156

* Fix comment alignment

* Fix date() to not reverse the array

This was only needed due to some malformed test combat lines which no
longer exists.

* Rewrite outdated comment

* Genericise `getSortedVideos()`

- Rename `getSortedVideos()` => `getSortedFiles()` and refactgor it
  to support any glob pattern, but still providing the sort mechanism.

- Rename `VideoInfo` => `FileInfo` type and move it to `types.ts`.

- Allow `CombatLogParser` to take a new option `fileFinderFn` which will
  provide it with a way to find files; this is mainly to avoid
  duplicating the code from `getSortedFiles()`.

  Incidentally, this move also made `getLatestLog()` async.

- Fix missing import of `Size` in `types.ts`

* Address comments in PR 156

* Remove 'Classic (Vanilla)'

It isn't and won't be supported.

* Add comment about `CombatLogParserOptionsType`

Also add more removal of Classic Vanilla WoW.