Fix breaking change to `splitLogLine()`
In an attempt to make TypeScript happy, all `continue` statements in the
block of `switch (char) {` were changed to `break` with a single
continue at the end.

This had the unfortunate consequence that it didn't parse anything
because it never added non-special characters to the `value` variable.

This fixes that.