Add methods for configurable resolution
Add `obsBaseResolution` and `obsOutputResolution` as configurable
settings.

Only `obsOutputResolution` can be configured at the moment, and base
resolution will always be set to the resolution of the selected monitor.

Both have been made configurable'ish because it is intended to be
exposed at a later time with a different mechanic that isn't suitable
for this change, and it was easier to add both now.