Move config schema and type to configSchema.ts
In order to avoid duplication of the configuration keys in
`useSettings.ts, the configuration schema and its associated type has
been moved to a new file that can be imported in the frontend free from
filesystem imports.

`configValues` in `useSettings.ts` has been moved outside of the
`useState()` method as we only need to load them from the configuration
store _once_ per config window creation as they cannot and should not
change outside of the config window.