Change ConfigService logging
We don't need ConfigService to log every single set/get indiscriminately
so this change introduces a kind of batching of the log output and
removal of the logging of `get()` as that seems rather unnecessary.

On instantiation, `ConfigService` will dump the current configuration to
the console as a baseline and when `set()` is called it'll batch up all
the changes within a certain amount of time and dump them to the
console afterwards to make it more useful and readable/non spammy.