Only record audio when buffer is recording
To avoid Windows not going to sleep because we keep a microphone
open, this change introduces a mechanism for only enabling the sound
sources when the buffer is recording, e.g. when Warcraft is running.

Instead of configuring sources when OBS is initialized, we'll postpone
that until OBS actually starts recording and clear the recording sources
when it ends.

It works like this:

Start:

- `obsRecorder.start()`
- Setup all sources

Stop:

- `obsRecorder.stop()`
- Clear and release all sources

Furthermore: simplify the filtering and adding of audio sources.